@@ -7,66 +7,68 @@
   "license": "Apache-2.0",
   "scripts": {
     "serve": "vue-cli-service serve",
-    "serve:fuseki": "nodemon src/services/mock/json-server.js src/services/mock/db.json",
-    "serve:offline": "concurrently --names 'SERVER,CLIENT' --prefix-colors 'yellow,blue' --success \"first\" --kill-others 'yarn run serve:fuseki' 'yarn run serve'",
     "build": "vue-cli-service build",
     "test:unit": "vue-cli-service test:unit",
     "test:e2e": "bash -c 'concurrently --names \"SERVER,TESTS\" --prefix-colors \"yellow,blue\" --success \"first\" --kill-others \"yarn run serve:fuseki\" \"vue-cli-service test:e2e ${0}\"'",
-    "coverage:unit": "nyc vue-cli-service test:unit",
+    "lint": "vue-cli-service lint",
     "coverage:e2e": "bash -c 'concurrently --names \"SERVER,TESTS\" --prefix-colors \"yellow,blue\" --success \"first\" --kill-others \"yarn run serve:fuseki\" \"nyc vue-cli-service test:e2e ${0}\"'",
-    "lint": "vue-cli-service lint"
+    "coverage:unit": "nyc vue-cli-service test:unit",
+    "serve:fuseki": "nodemon src/services/mock/json-server.js src/services/mock/db.json",
+    "serve:offline": "concurrently --names 'SERVER,CLIENT' --prefix-colors 'yellow,blue' --success \"first\" --kill-others 'yarn run serve:fuseki' 'yarn run serve'"
   },
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^6.1.1",
-    "@fortawesome/free-solid-svg-icons": "^6.1.1",
-    "@fortawesome/vue-fontawesome": "^2.0.6",
-    "@popperjs/core": "^2.11.5",
-    "@triply/yasqe": "^4.2.23",
-    "@triply/yasr": "^4.2.23",
-    "axios": "^0.26.1",
-    "bootstrap": "^5.2.0-beta1",
-    "core-js": "^3.21.1",
-    "follow-redirects": "^1.14.9",
-    "nanoid": "^3.3.1",
-    "vue": "^2.6.14",
-    "vue-router": "^3.5.3",
-    "vue-upload-component": "^2.8.22"
+    "@fortawesome/fontawesome-svg-core": "^6.2.0",
+    "@fortawesome/free-solid-svg-icons": "^6.2.0",
+    "@fortawesome/vue-fontawesome": "^3.0.1",
+    "@popperjs/core": "^2.11.6",
+    "@triply/yasqe": "^4.2.27",
+    "@triply/yasr": "^4.2.27",
+    "@vue/compat": "^3.2.39",
+    "axios": "^0.27.2",
+    "bootstrap": "^5.2.1",
+    "core-js": "^3.25.2",
+    "follow-redirects": "^1.15.2",
+    "mitt": "^3.0.0",
+    "nanoid": "^4.0.0",
+    "vue": "^3.2.39",
+    "vue-router": "^4.1.5",
+    "vue-upload-component": "^3.1.2"
   },
   "devDependencies": {
-    "@babel/register": "^7.17.0",
-    "@cypress/code-coverage": "^3.9.12",
-    "@cypress/vue": "^4.0.0",
-    "@cypress/webpack-preprocessor": "^5.11.0",
+    "@babel/eslint-parser": "^7.19.1",
+    "@babel/register": "^7.18.9",
+    "@cypress/code-coverage": "^3.10.0",
+    "@cypress/vue": "^4.2.0",
+    "@cypress/webpack-preprocessor": "^5.12.2",
     "@istanbuljs/nyc-config-babel": "^3.0.0",
-    "@vue/cli-plugin-babel": "^5.0.1",
-    "@vue/cli-plugin-e2e-cypress": "^5.0.5",
-    "@vue/cli-plugin-eslint": "^5.0.1",
-    "@vue/cli-plugin-unit-mocha": "~5.0.1",
-    "@vue/cli-service": "^5.0.1",
-    "@vue/eslint-config-standard": "^6.1.0",
-    "@vue/test-utils": "^1.3.0",
-    "babel-eslint": "^10.1.0",
+    "@vue/cli-plugin-babel": "~5.0.8",
+    "@vue/cli-plugin-e2e-cypress": "^5.0.8",
+    "@vue/cli-plugin-eslint": "~5.0.8",
+    "@vue/cli-plugin-unit-mocha": "~5.0.8",
+    "@vue/cli-service": "~5.0.8",
+    "@vue/compiler-sfc": "^3.2.39",
+    "@vue/eslint-config-standard": "^8.0.1",
+    "@vue/test-utils": "^2.0.2",
     "babel-plugin-istanbul": "^6.1.1",
     "chai": "^4.3.6",
-    "concurrently": "^7.0.0",
-    "cypress": "^10.3.0",
-    "eslint": "^7.22.0",
-    "eslint-plugin-import": "^2.25.4",
+    "concurrently": "^7.4.0",
+    "cypress": "^10.8.0",
+    "eslint": "^8.23.1",
+    "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-promise": "^6.0.0",
-    "eslint-plugin-standard": "^4.1.0",
-    "eslint-plugin-vue": "^8.4.0",
-    "express": "^4.17.2",
+    "eslint-plugin-promise": "^6.0.1",
+    "eslint-plugin-standard": "^5.0.0",
+    "eslint-plugin-vue": "^9.5.1",
+    "express": "^4.18.1",
     "istanbul-instrumenter-loader": "^3.0.1",
-    "istanbul-lib-coverage": "^3.0.0",
+    "istanbul-lib-coverage": "^3.2.0",
     "json-server": "^0.17.0",
-    "nodemon": "^2.0.19",
+    "nodemon": "^2.0.20",
     "nyc": "^15.1.0",
-    "sass": "^1.53.0",
-    "sass-loader": "^10.2.0",
+    "sass": "^1.54.9",
+    "sass-loader": "^13.0.2",
     "sinon": "^14.0.0",
-    "vue-template-compiler": "^2.6.14",
-    "webpack": "^5.73.0"
+    "webpack": "^5.74.0"
   },
   "overrides": {
     "terser": "^5.14.2"
