@@ -68,7 +68,7 @@ export default {
     }, 5000)
   },
 
-  beforeDestroy () {
+  beforeUnmount () {
     clearInterval(this.interval)
     this.interval = null
   }
