@@ -33,14 +33,19 @@
                     :items="items"
                     :is-busy="isBusy"
                   >
-                    <template v-slot:empty>
-                      <h4>No datasets created - <router-link to="/manage/new">add one</router-link></h4>
+                    <template #empty>
+                      <h4>
+                        No datasets created - <router-link to="/manage/new">
+                          add one
+                        </router-link>
+                      </h4>
                     </template>
-                    <template v-slot:cell(actions)="data">
+                    <template #cell(actions)="data">
                       <button
                         @click="$router.push(`/dataset${data.item.name}/query`)"
                         type="button"
-                        class="btn btn-primary me-0 me-md-2 mb-2 mb-md-0 d-block d-md-inline-block">
+                        class="btn btn-primary me-0 me-md-2 mb-2 mb-md-0 d-block d-md-inline-block"
+                      >
                         <FontAwesomeIcon icon="question-circle" />
                         <span class="ms-1">query</span>
                       </button>
@@ -51,16 +56,22 @@
                             Confirm
                           </div>
                           <div class="text-center">
-                            <div class="alert alert-danger">Are you sure you want to delete dataset {{ data.item.name }}?<br/><br/>This action cannot be reversed.</div>
+                            <div class="alert alert-danger">
+                              Are you sure you want to delete dataset {{ data.item.name }}?<br /><br />This action cannot be reversed.
+                            </div>
                             <button
                               @click="hidePopover();deleteDataset(data.item.name)"
                               class="btn btn-primary me-2"
-                            >submit</button>
+                            >
+                              submit
+                            </button>
                             <button
                               @click="hidePopover()"
                               type="button"
                               class="btn btn-secondary"
-                            >cancel</button>
+                            >
+                              cancel
+                            </button>
                           </div>
                         </div>
                       </div>
@@ -81,17 +92,23 @@
                             Confirm
                           </div>
                           <div class="text-center">
-                            <div class="alert alert-danger">Are you sure you want to create a backup of dataset {{ data.item.name }}?<br/><br/>This action may take some time to complete.</div>
+                            <div class="alert alert-danger">
+                              Are you sure you want to create a backup of dataset {{ data.item.name }}?<br /><br />This action may take some time to complete.
+                            </div>
                             <button
                               @click="hidePopover();backupDataset(data.item.name)"
                               type="button"
                               class="btn btn-primary me-2"
-                            >submit</button>
+                            >
+                              submit
+                            </button>
                             <button
                               @click="hidePopover()"
                               type="button"
                               class="btn btn-secondary"
-                            >cancel</button>
+                            >
+                              cancel
+                            </button>
                           </div>
                         </div>
                       </div>
@@ -109,14 +126,16 @@
                       <button
                         @click="$router.push(`/dataset${data.item.name}/upload`)"
                         type="button"
-                        class="btn btn-primary me-0 me-md-2 me-0 mb-2 mb-md-0 d-block d-md-inline-block">
+                        class="btn btn-primary me-0 me-md-2 me-0 mb-2 mb-md-0 d-block d-md-inline-block"
+                      >
                         <FontAwesomeIcon icon="upload" />
                         <span class="ms-1">add data</span>
                       </button>
                       <button
                         @click="$router.push(`/dataset${data.item.name}/info`)"
                         type="button"
-                        class="btn btn-primary me-0 mb-md-0 d-block d-md-inline-block">
+                        class="btn btn-primary me-0 mb-md-0 d-block d-md-inline-block"
+                      >
                         <FontAwesomeIcon icon="tachometer-alt" />
                         <span class="ms-1">info</span>
                       </button>
@@ -124,7 +143,7 @@
                   </table-listing>
                 </div>
               </div>
-           </div>
+            </div>
           </div>
         </div>
       </div>
@@ -135,6 +154,7 @@
 <script>
 import Menu from '@/components/manage/Menu'
 import listDatasets from '@/mixins/list-datasets'
+import listDatasetsNavigationGuards from '@/mixins/list-datasets-navigation-guards'
 import TableListing from '@/components/dataset/TableListing'
 import { library } from '@fortawesome/fontawesome-svg-core'
 import { faTimesCircle, faDownload, faTachometerAlt } from '@fortawesome/free-solid-svg-icons'
@@ -150,6 +170,8 @@ export default {
   mixins: [
     listDatasets
   ],
+
+  ...listDatasetsNavigationGuards,
 
   data () {
     return {
