@@ -16,29 +16,33 @@
  */
 
 import { expect } from 'chai'
-import { createLocalVue, shallowMount } from '@vue/test-utils'
+import { mount } from '@vue/test-utils'
 import ServerStatus from '@/components/ServerStatus.vue'
 import ServerStatusModel from '@/model/server.status'
 import sinon from 'sinon'
 
-const localVue = createLocalVue()
-
 let count = 0
-
-localVue.prototype.$fusekiService = {
-  async getServerStatus () {
-    count += 1
-    return new ServerStatusModel(true, `OK ${count}`)
-  }
-}
 
 describe('ServerStatus', () => {
   let fakeClock
   let component
   const mountFunction = options => {
-    return shallowMount(ServerStatus, {
-      localVue,
-      ...options
+    const mountOptions = Object.assign(options || {}, {
+      shallow: true,
+      global: {
+        mocks: {
+          $fusekiService: {
+            async getServerStatus () {
+              console.log('Hola!')
+              count += 1
+              return new ServerStatusModel(true, `OK ${count}`)
+            }
+          }
+        }
+      }
+    })
+    return mount(ServerStatus, {
+      ...mountOptions
     })
   }
   beforeEach(async () => {
@@ -58,7 +62,7 @@ describe('ServerStatus', () => {
   })
   it('destroys the interval', async () => {
     expect(component.vm.interval).to.not.equal(null)
-    component.vm.$destroy()
+    component.unmount()
     expect(component.vm.interval).to.equal(null)
   })
 })
